Create reservation test case
